fix(review): address PR #1319 review comments
- Add debug log to auth/check endpoint
- Replace hardcoded password validation with dynamic securitySettings
- Extract PasswordPolicyHint to shared common component
- Add password change UI to UserSettingsModal
- Fix prettier formatting issues
