+++ frontend/src/components/features/ForceChangePasswordModal.tsx
@@ -6,7 +6,7 @@
  */
 
 import React, { useState } from 'react';
-import { Modal, Button, TextInput } from '@/components/common';
+import { Modal, Button, TextInput, PasswordPolicyHint } from '@/components/common';
 import { useAuth, useLanguage, useMustChangePassword, useSecuritySettings } from '@/hooks';
 import { t } from '@/i18n';
 
@@ -26,30 +26,6 @@ export const ForceChangePasswordModal: React.FC = () => {
     setSkipped(true);
   };
 
-  // Password policy hint component
-  const PasswordPolicyHint = () => {
-    const policy = securitySettings;
-    if (!policy) return null;
-
-    const requirements: string[] = [];
-    requirements.push(`${t('passwordMinLength', language)}: ${policy.password_min_length ?? 8}`);
-    if (policy.password_require_uppercase) requirements.push(t('requireUppercase', language));
-    if (policy.password_require_lowercase) requirements.push(t('requireLowercase', language));
-    if (policy.password_require_number) requirements.push(t('requireNumber', language));
-    if (policy.password_require_special) requirements.push(t('requireSpecial', language));
-
-    return (
-      <div className="password-policy-hint text-muted small mt-1">
-        <div>{t('passwordRequirements', language)}:</div>
-        <ul className="mb-0 ps-3" style={{ fontSize: '0.85em' }}>
-          {requirements.map((req, idx) => (
-            <li key={idx}>{req}</li>
-          ))}
-        </ul>
-      </div>
-    );
-  };
-
   const handleSubmit = async () => {
     setError(null);
 
@@ -63,8 +39,11 @@ export const ForceChangePasswordModal: React.FC = () => {
       return;
     }
 
-    if (newPassword.length < 8) {
-      setError(t('passwordTooShort', language) ?? 'Password must be at least 8 characters');
+    const minLength = securitySettings?.password_min_length ?? 8;
+    if (newPassword.length < minLength) {
+      setError(
+        t('passwordTooShort', language) ?? `Password must be at least ${minLength} characters`
+      );
       return;
     }
